[feat] Booth 모듈 네비게이션 연동
Nested Navigation 구현, Map 화면과 연결, BoothId 를 전달

+++ feature/map/src/main/kotlin/com/unifest/android/feature/map/navigation/MapNavigation.kt
@@ -15,7 +15,7 @@ fun NavController.navigateToMap(navOptions: NavOptions) {
 
 fun NavGraphBuilder.mapNavGraph(
     padding: PaddingValues,
-    onNavigateToBooth: () -> Unit,
+    onNavigateToBooth: (Long) -> Unit,
 ) {
     composable(route = MAP_ROUTE) {
         MapRoute(
